eski sürümler/v2 correct/ai gez/core_app/requirements.txt
selenium
webdriver-manager
seleniumbase
requests
beautifulsoup4
curl_cffi
python-dotenv
google-cloud-storage

# Telegram
python-telegram-bot==21.1.1
pyvirtualdisplay==3.0

# Optional runtime for WARP + DrissionPage flow
DrissionPage

nodriver
